requirements.txt
requests~=2.28.2
tqdm~=4.64.1
pandas~=1.5.3
retry~=0.9.2
multitasking~=0.0.11
jsonpath~=0.82
rich~=13.3.1
numpy~=1.24.2
mysql-connector-python~=8.0.32
setuptools~=65.5.1
django-scheduler~=0.10.1